Initial release
I'm not going to make an individual commit for everything I did from the other alpha version, I'm going to test this in the following OS:
- RHEL/CentOS 8
- Manjaro/Arch kernel 5.10 (development computer)
-Ubuntu LTS/Debian 10
- Windows 10, with and without WSL.

After making sure that the script, runs in these systems, if it even runs, I will have an idea on how to properly implement the rest of this script into a package that can be downloaded/ installed with a bash script.
Please read the comments inside the file to understand why it works the way it works, and the readme for learning how to use lzh, I ain't even talking to the internet, this is for the future me, that probably will no longer remember most of this logic.